# Minimize and speed up the Collatz conjecture bench