@ekzhang 2022-03-02
Nice!! Cool to see that in both cases, the asymptotic time complexity appears to be the same (all are around `-50%`). I think this means the semi-naive evaluation is probably working.